requirements.txt
amqp==5.3.1
billiard==4.2.1
celery==5.5.2
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.2
greenlet==3.2.2
kombu==5.5.3
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyMuPDF==1.25.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.1.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
